fix some functions

+++ srcs/vec/from.c
@@ -6,7 +6,7 @@
 /*   By: cchudant <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/26 14:13:16 by cchudant          #+#    #+#             */
-/*   Updated: 2019/10/26 14:23:02 by cchudant         ###   ########.fr       */
+/*   Updated: 2019/11/15 06:21:18 by cchudant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@ t_vec2d			vec2d_from(t_vec2i v)
 {
 	t_vec2d r;
 
-	r.x = v.x;
-	r.y = v.y;
+	r.x = (double)v.x;
+	r.y = (double)v.y;
 	return (r);
 }
 
@@ -25,8 +25,8 @@ t_vec2i			vec2i_from(t_vec2d v)
 {
 	t_vec2i r;
 
-	r.x = v.x;
-	r.y = v.y;
+	r.x = (int)v.x;
+	r.y = (int)v.y;
 	return (r);
 }
 
@@ -34,9 +34,9 @@ t_vec3d			vec3d_from(t_vec3i v)
 {
 	t_vec3d r;
 
-	r.x = v.x;
-	r.y = v.y;
-	r.z = v.z;
+	r.x = (double)v.x;
+	r.y = (double)v.y;
+	r.z = (double)v.z;
 	return (r);
 }
 
@@ -44,8 +44,8 @@ t_vec3i			vec3i_from(t_vec3d v)
 {
 	t_vec3i r;
 
-	r.x = v.x;
-	r.y = v.y;
-	r.z = v.z;
+	r.x = (int)v.x;
+	r.y = (int)v.y;
+	r.z = (int)v.z;
 	return (r);
 }
